Break up mutations and initalState into separate modules
