+++ kolibri/plugins/learn/assets/src/state/initialState.js
@@ -1,0 +1,15 @@
+import { PageNames } from '../constants';
+
+export default {
+  pageName: PageNames.TOPICS_CHANNEL,
+  pageState: {
+    topics: [],
+    contents: [],
+    searchTerm: '',
+  },
+  learnAppState: {
+    memberships: [],
+  },
+  examLog: {},
+  examAttemptLogs: {},
+};
